final changes

+++ manageUsers/deleteUser.py
@@ -4,12 +4,14 @@ from fastapi import HTTPException
 from http import HTTPStatus
 
 def users_delete(cur_user,id: int):
+    role = databaseHelper.role_power(cur_user.id)
+    if role == 'User':
+        raise HTTPException(status_code = HTTPStatus.UNAUTHORIZED, detail = logError.USER_IS_UNAUTHORISED)
     if id == 1:
         raise HTTPException(status_code = HTTPStatus.UNAUTHORIZED, detail = logError.USER_IS_UNAUTHORISED)
     to_delete = databaseHelper.list_user_id(id)
     if not to_delete:
         raise HTTPException(status_code = HTTPStatus.NOT_FOUND, detail = logError.USER_NOT_FOUND)
-    role = databaseHelper.role_power(cur_user.id)
     if role == 'Superadmin':
         return databaseHelper.delete_user_superadmin(id)
     elif role == 'Admin':
